+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add metric `irsa_operator_acm_certificate_not_after` metric to expose the `NotAfter` timestamp of the ACM certificate.
+
+### Changed
+
+- Add a cache to ACM service to avoid hitting the API too hard.
+
+### Fixed
+
+- Vintage: fix not performing validation on renewal of certificate.
+
 ## [0.26.3] - 2024-04-10
 
 ### Fixed